requirements.txt
streamlit
bcrypt
whisper
PyMuPDF
faiss-cpu
sentence-transformers
scikit-learn
yt_dlp
python-dotenv
google-generativeai
